Remove old ToDo and rename Deadline as ToDo
