+++ src/main/java/seedu/calidr/model/ModelManager.java
@@ -14,7 +14,6 @@ import javafx.collections.transformation.FilteredList;
 import seedu.calidr.commons.core.GuiSettings;
 import seedu.calidr.commons.core.LogsCenter;
 import seedu.calidr.model.person.Person;
-import seedu.calidr.model.task.Deadline;
 import seedu.calidr.model.task.Event;
 import seedu.calidr.model.task.Task;
 import seedu.calidr.model.task.ToDo;
@@ -138,8 +137,7 @@ public class ModelManager implements Model {
     public ArrayList<Task> getTaskList() {
         // TODO remove stub
         ArrayList<Task> taskList = new ArrayList<>();
-        taskList.add(new ToDo("CS2103T"));
-        taskList.add(new Deadline("CS2101", LocalDateTime.of(2023, 3, 5, 3, 10)));
+        taskList.add(new ToDo("CS2101", LocalDateTime.of(2023, 3, 5, 3, 10)));
         taskList.add(new Event("CS3211", LocalDateTime.of(2023, 3, 6, 10, 10), LocalDateTime.of(2023, 3, 7, 10, 10)));
         return taskList;
     }
