+++ src/test/java/seedu/calidr/testutil/TypicalTasks.java
@@ -2,7 +2,6 @@ package seedu.calidr.testutil;
 
 import java.time.LocalDateTime;
 
-import seedu.calidr.model.task.Deadline;
 import seedu.calidr.model.task.Event;
 import seedu.calidr.model.task.Priority;
 import seedu.calidr.model.task.ToDo;
@@ -12,15 +11,11 @@ import seedu.calidr.model.task.ToDo;
  */
 public class TypicalTasks {
 
-    public static final ToDo TODO1 = new ToDo("CS2103T Quiz 1");
-    public static final ToDo TODO2 = new ToDo("CS2103T Quiz 2", Priority.HIGH);
-    public static final ToDo TODO3 = new ToDo("CS2103T Quiz 3", Priority.LOW);
-
     public static final LocalDateTime DATE_TIME1 = LocalDateTime.of(2019, 10, 10, 10, 10);
     public static final LocalDateTime DATE_TIME2 = LocalDateTime.of(2019, 10, 11, 10, 10);
-    public static final Deadline DEADLINE1 = new Deadline("CS2101 Quiz 2", DATE_TIME1, Priority.HIGH);
+    public static final ToDo TODO1 = new ToDo("CS2101 Quiz 2", DATE_TIME1, Priority.HIGH);
 
-    public static final Deadline DEADLINE2 = new Deadline("CS2101 Quiz 3", DATE_TIME2, Priority.LOW);
+    public static final ToDo TODO2 = new ToDo("CS2101 Quiz 3", DATE_TIME2, Priority.LOW);
 
     public static final Event EVENT1 = new Event("CS2103T Lecture 1", DATE_TIME1, DATE_TIME2);
